Moved op_test_utils
